refactor: thumbnail 크롤링 하여 저장하는 메서드는 비동기로 일어나도록 수정
